requirements.txt
llama-index
llama-index-vector-stores-supabase
llama-index-vector-stores-postgres
llama-index-embeddings-openai
llama-index-llms-openai
fastapi
uvicorn
supabase
python-dotenv
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
asyncpg
llama-index-storage-chat-store-postgres
psycopg[binary]